remove block manager and move to block
